Рабил JS на модули
