+++ index.html
@@ -164,6 +164,8 @@
       </div>
     </section>
   </template>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
+<script src="js/data.js" type="module"></script>
+<script src="js/util.js" type="module"></script>
 </body>
 </html>
